# Speed up specs by 9.5 minutes