--- a/core/spec/models/spree/exchange_spec.rb
+++ b/core/spec/models/spree/exchange_spec.rb
@@ -30,9 +30,13 @@ module Spree
     end
 
     describe "#perform!" do
-      let(:return_item) { create(:exchange_return_item) }
+      let(:return_item) { create(:exchange_return_item, inventory_unit: inventory_unit) }
       let(:return_items) { [return_item] }
-      let(:order) { return_item.return_authorization.order }
+
+      let(:inventory_unit) { create(:inventory_unit, order: order, line_item: line_item) }
+      let(:order) { create(:shipped_order, line_items_count: 1) }
+      let(:line_item) { order.line_items.first }
+
       subject { exchange.perform! }
       before { return_item.exchange_variant.stock_items.first.adjust_count_on_hand(20) }