--- a/core/spec/models/spree/order_spec.rb
+++ b/core/spec/models/spree/order_spec.rb
@@ -745,10 +745,10 @@ describe Spree::Order do
 
   describe '#quantity' do
     # Uses a persisted record, as the quantity is retrieved via a DB count
-    let(:order) { create :order_with_line_items }
+    let(:order) { create :order_with_line_items, line_items_count: 2 }
 
     it 'sums the quantity of all line items' do
-      expect(order.quantity).to eq 5
+      expect(order.quantity).to eq 2
     end
   end
 
@@ -760,15 +760,6 @@ describe Spree::Order do
       ]
 
       expect(subject.pre_tax_item_amount).to eq 19.0
-    end
-  end
-
-  describe '#quantity' do
-    # Uses a persisted record, as the quantity is retrieved via a DB count
-    let(:order) { create :order_with_line_items }
-
-    it 'sums the quantity of all line items' do
-      expect(order.quantity).to eq 5
     end
   end