I went looking for some low-hanging fruit to speed up specs.  I found some. 

All these times are on my own computer.

Overall: 9.5min speedup (when run serially)

Core (7.5min speedup)
  - Before: 762s
  - After:  306s

Backend (92s speedup)
  - Before: 473s
  - After:  381s

API (26s speedup)
  - Before: 93s
  - After:  67s

Frontend (no change)
  - Before: 123s
  - After:  121s