@jordan-brough 2015-04-10
Speedup on CircleCI was about 9 minutes:

![screenshot-2015-04-10-08-47-34](https://cloud.githubusercontent.com/assets/23050/7090198/7c85f4e4-df5e-11e4-9648-b5a2d2aa5327.jpg)

(most recent time is this one)