@athal7 2015-04-10
:+1: I'd be interested to try to make the line_item_count=1 by default at some point. This is a great step towards that.