@jordan-brough 2015-04-10
@athal This actually does do that. :)  You mean [this](https://github.com/bonobos/spree/pull/327/files#diff-b88375f378b6372af0d19d209f45d104L24) right?  There were a couple places where I still explicitly specified it as 1 since it seemed like the spec specifically expected 1 line item to be present.