@cbrunsdon 2015-04-10
Ha, excellent. This was on my "todo" list for solidusconf.

:+1: